Shuffle and consolidate tests
Prefer unit tests instead of integration tests
